+++ src/Category_Colors/Admin.php
@@ -20,7 +20,7 @@ class Admin {
 		add_action( 'admin_notices', array( $this, 'plugin_fail_msg' ) );
 		add_action( 'tribe_settings_below_tabs_tab_category-colors', array( $this, 'is_saved' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'load_teccc_js_css' ) );
-		load_plugin_textdomain( 'events-calendar-category-colors', false, TECCC_LANG );
+		load_plugin_textdomain( 'the-events-calendar-category-colors', false, TECCC_LANG );
 	}
 
 
@@ -33,12 +33,12 @@ class Admin {
 		if ( current_user_can( 'activate_plugins' ) && is_admin() ) {
 			if ( ! class_exists( 'Tribe__Events__Events' ) ) {
 				$url   = 'plugin-install.php?tab=plugin-information&plugin=the-events-calendar&TB_iframe=true';
-				$title = __( 'The Events Calendar', 'events-calendar-category-colors' );
-				echo '<div class="error"><p>'.sprintf( __( 'To begin using The Events Calendar Category Colors, please install the latest version of <a href="%s" class="thickbox" title="%s">The Events Calendar</a>.', 'events-calendar-category-colors' ),$url, $title ).'</p></div>';
+				$title = __( 'The Events Calendar', 'the-events-calendar-category-colors' );
+				echo '<div class="error"><p>'.sprintf( __( 'To begin using The Events Calendar Category Colors, please install the latest version of <a href="%s" class="thickbox" title="%s">The Events Calendar</a>.', 'the-events-calendar-category-colors' ),$url, $title ).'</p></div>';
 			} elseif ( version_compare( Tribe__Events__Events::VERSION, '3.0', 'lt') ) {
 				$url   = 'plugin-install.php?tab=plugin-information&plugin=the-events-calendar&TB_iframe=true';
-				$title = __( 'The Events Calendar', 'events-calendar-category-colors' );
-				echo '<div class="error"><p>'.sprintf( __( 'You have The Events Calendar v.' . Tribe__Events__Events::VERSION . '. To begin using The Events Calendar Category Colors, please install the latest version of <a href="%s" class="thickbox" title="%s">The Events Calendar</a>.', 'events-calendar-category-colors' ),$url, $title ).'</p></div>';
+				$title = __( 'The Events Calendar', 'the-events-calendar-category-colors' );
+				echo '<div class="error"><p>'.sprintf( __( 'You have The Events Calendar v.' . Tribe__Events__Events::VERSION . '. To begin using The Events Calendar Category Colors, please install the latest version of <a href="%s" class="thickbox" title="%s">The Events Calendar</a>.', 'the-events-calendar-category-colors' ),$url, $title ).'</p></div>';
 			}
 		}
 	}
@@ -96,7 +96,7 @@ class Admin {
 		$categoryColorsTab = $this->teccc->load_config( 'admintab' );
 		add_action( 'tribe_settings_form_element_tab_category-colors', array( $this, 'form_header' ) );
 		add_action( 'tribe_settings_before_content_tab_category-colors', array( $this, 'settings_fields' ) );
-		new Tribe__Events__Settings_Tab( self::TAB_NAME, __( 'Category Colors', 'events-calendar-category-colors' ), $categoryColorsTab );
+		new Tribe__Events__Settings_Tab( self::TAB_NAME, __( 'Category Colors', 'the-events-calendar-category-colors' ), $categoryColorsTab );
 	}